+++ src/dakara_feeder/song.py
@@ -1,6 +1,10 @@
 import logging
 
-from dakara_feeder.metadata_parser import FFProbeMetadataParser, MediaParseError
+from dakara_feeder.metadata_parser import (
+    FFProbeMetadataParser,
+    MediaParseError,
+    NullMetadataParser,
+)
 from dakara_feeder.subtitle_parser import Pysubs2SubtitleParser, SubtitleParseError
 
 
@@ -16,15 +20,16 @@ class BaseSong:
     The main entry point of the class when used by the feeder is the
     `get_representation` method, that will call all the methods to get the song
     data:
-        - get_title;
-        - get_duration;
-        - get_version;
-        - get_detail;
-        - get_detail_video;
-        - get_tags;
-        - get_artists;
-        - get_works;
-        - get_lyrics.
+        - `get_title`;
+        - `get_duration`;
+        - `get_has_instrumental`;
+        - `get_version`;
+        - `get_detail`;
+        - `get_detail_video`;
+        - `get_tags`;
+        - `get_artists`;
+        - `get_works`;
+        - `get_lyrics`.
 
     You should override those methods to suit your needs. See the documentation
     of each method to learn what data format the must return.
@@ -32,33 +37,72 @@ class BaseSong:
     When calling `get_representation`, two special methods are also called for
     performing custom actions, the first one just on entering
     `get_representation`, and the other just befor leaving it:
-        - pre_process;
-        - post_process.
+        - `pre_process`;
+        - `post_process`.
 
     You should override those two methods as well. Typically, `pre_process`
     should be overriden to perform preparative actions which result would be
     used by the `get_` methods. On the other hand, `post_process` should be
     overriden to perform final actions on the representation.
 
+    Metadata of the video file are extracted using a metadata parser and stored
+    in the `metadata` attribute. The metadata parser to chose is decided by
+    setting the class attribute `metadata_parser_class`. The class must
+    implement `dakara_feeder.metadata_parser.MetadataParser` base class. So far,
+    two implemenations are available in the project:
+        - `dakara_feeder.metadata_parser.FFProbeMetadataParser`, based on
+            FFProbe, part of FFMpeg (external dependency). This is the recommended
+            and the default parser;
+        - `dakara_feeder.metadata_parser.MediainfoMetadataParser`, based on
+            MediaInfo (external dependency). Slower, may not work on Windows.
+
+    Metadata are available when calling `pre_process`.
+
+    If the metadata cannot be extracted from the video file for any reason, the
+    `metadata` attribute will contain a
+    `dakara_feeder.metadata_parser.NullMetadataParser` that always return null
+    values (e.g. 0 seconds duration).
+
     Args.
         base_directory (path.Path): path to the scanned directory.
         paths (directory_lister.SongPaths): paths of the song file.
 
     Attributes:
+        metadata_parser_class (type): Class of the metadata parser to use.
+            Default to `dakara_feeder.metadata_parser.FFProbeMetadataParser`.
         base_directory (path.Path): path to the scanned directory.
         video_path (path.Path): path to the song file, relative to the base
+            directory.
+        audio_path (path.Path): path to the audio file, relative to the base
             directory.
         sublitle_path (path.Path): path to the subtitle file, relative to the
             base directory.
         others_path (list of path.Path): list of paths to the other files,
             relative to the base directory.
+        metadata (dakara_feeder.metadata_parser.MetadataParser): Object for
+            containing metadata of the video file.
     """
+
+    metadata_parser_class = FFProbeMetadataParser
 
     def __init__(self, base_directory, paths):
         self.base_directory = base_directory
         self.video_path = paths.video
+        self.audio_path = paths.audio
         self.subtitle_path = paths.subtitle
         self.others_path = paths.others
+        self.metadata = NullMetadataParser.parse(self.video_path)
+
+    def parse_metadata(self):
+        """Use the requested metadata parser to parse video file
+        """
+        try:
+            self.metadata = self.metadata_parser_class.parse(
+                self.base_directory / self.video_path
+            )
+
+        except MediaParseError as error:
+            logger.error("Cannot parse metadata: {}".format(error))
 
     def pre_process(self):
         """Process preparative actions
@@ -101,26 +145,26 @@ class BaseSong:
         This method may be overriden. By default it returns the duration of the
         video file using FFProbe.
 
-        Duration can be extracted from the video file using a parser. Two
-        parsers are available in the project:
-            - `dakara_feeder.metadata_parser.FFProbeMetadataParser`, based on
-                FFProbe, part of FFMpeg (external dependency). This is the
-                recommended parser;
-            - `dakara_feeder.metadata_parser.MediainfoMetadataParser`, based on
-                MediaInfo (external dependency). Slower, does not work on
-                Windows.
 
         Returns:
             float: Duration of the song in seconds.
         """
-        try:
-            parser = FFProbeMetadataParser.parse(self.base_directory / self.video_path)
-            return parser.get_duration().total_seconds()
+        return self.metadata.get_duration().total_seconds()
 
-        except MediaParseError as error:
-            logger.error("Duration not parsed: {}".format(error))
-            # Set duration to zero when duration can't be extracted
-            return 0
+    def get_has_instrumental(self):
+        """Get the flag if the song has an instrumental track
+
+        Returns:
+            bool: True either if there is an extra audio file siding with the
+            video file, or if the video file has more than 2 audio tracks.
+        """
+        if self.audio_path:
+            return True
+
+        if self.metadata.get_audio_tracks_count() >= 2:
+            return True
+
+        return False
 
     def get_artists(self):
         """Get the list of artists
@@ -253,12 +297,14 @@ class BaseSong:
         Returns:
             dict: JSON-compiliant structure representing the song.
         """
+        self.parse_metadata()
         self.pre_process()
         representation = {
             "title": self.get_title(),
             "filename": str(self.video_path.basename()),
             "directory": str(self.video_path.dirname()),
             "duration": self.get_duration(),
+            "has_instrumental": self.get_has_instrumental(),
             "version": self.get_version(),
             "detail": self.get_detail(),
             "detail_video": self.get_detail_video(),