+++ src/dakara_feeder/subtitle_parser.py
@@ -1,9 +1,19 @@
 import re
 from abc import ABC, abstractmethod
-from collections import OrderedDict
 
 import pysubs2
 from dakara_base.exceptions import DakaraError
+
+
+def is_subtitle(filename):
+    """Check if the provided file is a subtitle
+
+    Check the admissible file extensions for pysubs2.
+
+    Returns:
+        bool: True if the filename is a subtitle.
+    """
+    return filename.ext in pysubs2.formats.FILE_EXTENSION_TO_FORMAT_IDENTIFIER
 
 
 class SubtitleParser(ABC):
@@ -29,6 +39,18 @@ class SubtitleParser(ABC):
             SubtitleParser: instance of the class for the given file.
         """
 
+    @classmethod
+    @abstractmethod
+    def parse_string(cls, filecontent):
+        """Read a subtitle stream and store the lyrics
+
+        Args:
+            filecontent (str): content of the file to extract lyrics from.
+
+        Returns:
+            SubtitleParser: instance of the class for the given content.
+        """
+
     @abstractmethod
     def get_lyrics(self):
         """Extract lyrics
@@ -39,7 +61,7 @@ class SubtitleParser(ABC):
 
 
 class TXTSubtitleParser(SubtitleParser):
-    """Subtitle parser for txt files
+    """Subtitle parser for plain txt files
 
     >>> from Path import path
     >>> file_path = Path("path/to/file")
@@ -62,6 +84,18 @@ class TXTSubtitleParser(SubtitleParser):
             TXTSubtitleParser: instance of the class for the given file.
         """
         return cls(filepath.text())
+
+    @classmethod
+    def parse_string(cls, filecontent):
+        """Read a subtitle file and store the lyrics
+
+        Args:
+            filecontent (str): content of the file to extract lyrics from.
+
+        Returns:
+            SubtitleParser: instance of the class for the given content.
+        """
+        return cls(filecontent)
 
     def get_lyrics(self):
         """Extract lyrics
@@ -122,19 +156,35 @@ class Pysubs2SubtitleParser(SubtitleParser):
             Pysubs2SubtitleParser: instance of the class for the given file.
         """
         try:
-            instance = cls(pysubs2.load(filepath))
+            return cls(pysubs2.load(filepath))
 
         except FileNotFoundError as error:
             raise SubtitleNotFoundError(
-                "Subtitle file {} not found".format(filepath)
+                "Subtitle file '{}' not found".format(filepath)
             ) from error
 
-        except BaseException as error:
+        except Exception as error:
             raise SubtitleParseError(
-                "Error when parsing subtitle file {}: {}".format(filepath, error)
+                "Error when parsing subtitle file '{}': {}".format(filepath, error)
             ) from error
 
-        return instance
+    @classmethod
+    def parse_string(cls, filecontent):
+        """Read a subtitle file and store the lyrics
+
+        Args:
+            filecontent (str): content of the file to extract lyrics from.
+
+        Returns:
+            SubtitleParser: instance of the class for the given content.
+        """
+        try:
+            return cls(pysubs2.SSAFile.from_string(filecontent))
+
+        except Exception as error:
+            raise SubtitleParseError(
+                "Error when parsing subtitle content: {}".format(error)
+            ) from error
 
     def get_lyrics(self):
         """Gives the cleaned text of the Event block
@@ -184,16 +234,6 @@ class Pysubs2SubtitleParser(SubtitleParser):
             event_previous = event
 
         return "\n".join(lyrics)
-
-
-PARSER_BY_EXTENSION = OrderedDict(
-    (
-        (".ass", Pysubs2SubtitleParser),
-        (".ssa", Pysubs2SubtitleParser),
-        (".srt", Pysubs2SubtitleParser),
-        (".txt", TXTSubtitleParser),
-    )
-)
 
 
 class SubtitleParseError(DakaraError):